+++ src/explorer/issues_treeview.ts
@@ -29,6 +29,9 @@ export class TfsecIssueProvider implements vscode.TreeDataProvider<TfsecTreeItem
 			}
 			// create the file watcher to refresh the tree when changes are made
 			fs.watch(this.resultsStoragePath, (eventType, filename) => {
+				if (eventType !== "change") {
+					return;
+				}
 				vscode.window.showInformationMessage("tfsec run complete, results file updated");
 				// short wait for the file to be written before refreshing the tree
 				setTimeout(() => { vscode.commands.executeCommand("tfsec.refresh"); }, 250);